pyproject.toml
[tool.poetry]
name = "amanogawa"
version = "0.1.0"
description = ""
authors = ["Jonathan Crum <jcrum@theobogroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
devtools = "^0.12.2"
fastapi = "^0.111.0"
google = "^3.0.0"
google-api-python-client = "^2.135.0"
google-cloud-iam = "^2.15.0"
google-cloud-language = "^2.13.3"
lxml = "^5.2.2"
nltk = "^3.8.1"
pandas = "^2.2.2"
pendulum = "^3.0.0"
polars = "^0.20.31"
psycopg2 = "^2.9.9"
pygtrie = "^2.5.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"
requests = "^2.32.3"
sqlalchemy = "^2.0.31"
sudachidict-core = "^20240409"
sudachipy = ">=0.6.8"
pyarrow = "^16.1.0"
google-cloud-translate = "^3.15.3"
openai = "^1.35.7"
pywin32 = "^306"
voicevox-client = "^0.4.1"
pygame = "^2.6.0"
sudachidict-full = "^20240409"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "F401",   # module imported but unused
    "F403",   # 'from module import *' unused; unable to detect undefined names
    "F405",   # name may be undefined, or defined from star imports
    "F841",   # local variable 'remote' is assigned to but never used
]
